v.2.2.9 (#33)

+++ intaro.retailcrm/classes/general/history/RetailCrmHistory_v4.php
@@ -614,7 +614,7 @@ class RetailCrmHistory
                                         'CURRENCY' => \Bitrix\Currency\CurrencyManager::getBaseCurrency(),
                                         'LID' => $site,
                                         'BASE_PRICE' => $product['initialPrice'],
-                                        'NAME' => $product['name'] ? RCrmActions::fromJSON($product['name']) : $elem['NAME'],
+                                        'NAME' => $product['offer']['name'] ? RCrmActions::fromJSON($product['offer']['name']) : $elem['NAME'],
                                         'DETAIL_PAGE_URL' => $elem['URL']
                                     ));
                                 } else {
@@ -1186,16 +1186,16 @@ class RetailCrmHistory
         return $outputArray;
     }
     
-    public static function setProp($obj, $value = '', $prop)
+    public static function setProp($obj, $value = '', $prop = '')
     {
         if (!isset($obj)) {
             return false;
         }
-        if ($prop) {
+        if ($prop && $value) {
             $obj->setField($prop, $value);
-        } elseif ($value) {
+        } elseif ($value && !$prop) {
             $obj->setValue($value);
-        } else {
+        } elseif (!$value && !$prop) {
             $obj->delete();
         }
